Move startIteration to helpers.go
